+++ src/SolarFlareDriver.cc
@@ -532,11 +532,11 @@ SolarFlareDriver::newAddress(const ServiceLocator& serviceLocator) {
 
 /// See docs in the ``Driver'' class.
 void
-SolarFlareDriver::sendPacket(const Driver::Address *recipient,
-                             const void *header,
+SolarFlareDriver::sendPacket(const Driver::Address* recipient,
+                             const void* header,
                              uint32_t headerLen,
-                             Buffer::Iterator *payload,
-                             uint8_t priority)
+                             Buffer::Iterator* payload,
+                             int priority)
 {
 
     uint32_t udpPayloadLen = downCast<uint32_t>(headerLen
