+++ src/SolarFlareDriver.h
@@ -54,11 +54,11 @@ class SolarFlareDriver : public Driver {
     virtual void receivePackets(int maxPackets,
             std::vector<Received>* receivedPackets);
     virtual void release(char* payload);
-    virtual void sendPacket(const Driver::Address *recipient,
-                            const void *header,
+    virtual void sendPacket(const Driver::Address* recipient,
+                            const void* header,
                             uint32_t headerLen,
-                            Buffer::Iterator *payload,
-                            uint8_t priority = 0);
+                            Buffer::Iterator* payload,
+                            int priority = 0);
     virtual string getServiceLocator();
     virtual Driver::Address* newAddress(const ServiceLocator& serviceLocator);
